LUCENE-9558: Clean up package name conflicts for analyzers-icu.
