+++ lucene/CHANGES.txt
@@ -12,9 +12,6 @@ System Requirements
 
 API Changes
 
-* LUCENE-9317: Clean up package name conflicts between core and analyzers-common.
-  See MIGRATE.md for details. (David Ryan, Tomoko Uchida, Uwe Schindler, Dawid Weiss)
-
 * LUCENE-8474: RAMDirectory and associated deprecated classes have been
   removed. (Dawid Weiss)
 
@@ -67,6 +64,9 @@ API Changes
 * LUCENE-9516: Removed the ability to replace the IndexingChain / DocConsumer
   in Lucenes IndexWriter. The interface is not sufficient to efficiently
   replace the functionality with reasonable efforts. (Simon Willnauer)
+
+* LUCENE-9317 LUCENE-9558: Clean up package name conflicts between modules.
+  See MIGRATE.md for details. (David Ryan, Tomoko Uchida, Uwe Schindler, Dawid Weiss)
 
 Improvements
