+++ lucene/analysis/icu/src/java/org/apache/lucene/analysis/icu/ICUCollationKeyAnalyzer.java
@@ -14,19 +14,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.collation;
+package org.apache.lucene.analysis.icu;
 
 
-import org.apache.lucene.analysis.TokenStream;
-import org.apache.lucene.collation.tokenattributes.ICUCollatedTermAttributeImpl;
-import org.apache.lucene.util.AttributeFactory;
 
 import com.ibm.icu.text.Collator;
+import org.apache.lucene.analysis.Analyzer;
+import org.apache.lucene.analysis.core.KeywordTokenizer;
+import org.apache.lucene.collation.CollationKeyAnalyzer;
 
 /**
  * <p>
- *   Converts each token into its {@link com.ibm.icu.text.CollationKey}, and
- *   then encodes bytes as an index term.
+ *   Configures {@link KeywordTokenizer} with {@link ICUCollationAttributeFactory}.
+ * <p>
+ *   Converts the token into its {@link com.ibm.icu.text.CollationKey}, and
+ *   then encodes the CollationKey directly to allow it to
+ *   be stored as an index term.
  * </p>
  * <p>
  *   <strong>WARNING:</strong> Make sure you use exactly the same Collator at
@@ -48,44 +51,36 @@ import com.ibm.icu.text.Collator;
  * <p>
  *   CollationKeys generated by ICU Collators are not compatible with those
  *   generated by java.text.Collators.  Specifically, if you use 
- *   ICUCollationAttributeFactory to generate index terms, do not use 
- *   {@link CollationAttributeFactory} on the query side, or vice versa.
+ *   ICUCollationKeyAnalyzer to generate index terms, do not use 
+ *   {@link CollationKeyAnalyzer} on the query side, or vice versa.
  * </p>
  * <p>
- *   ICUCollationAttributeFactory is significantly faster and generates significantly
- *   shorter keys than CollationAttributeFactory.  See
+ *   ICUCollationKeyAnalyzer is significantly faster and generates significantly
+ *   shorter keys than CollationKeyAnalyzer.  See
  *   <a href="http://site.icu-project.org/charts/collation-icu4j-sun"
  *   >http://site.icu-project.org/charts/collation-icu4j-sun</a> for key
  *   generation timing and key length comparisons between ICU4J and
  *   java.text.Collator over several languages.
  * </p>
+ *
+ * @since 3.1
  */
-public class ICUCollationAttributeFactory extends AttributeFactory.StaticImplementationAttributeFactory<ICUCollatedTermAttributeImpl> {
-  private final Collator collator;
-  
+public final class ICUCollationKeyAnalyzer extends Analyzer {
+  private final ICUCollationAttributeFactory factory;
+
   /**
-   * Create an ICUCollationAttributeFactory, using 
-   * {@link TokenStream#DEFAULT_TOKEN_ATTRIBUTE_FACTORY} as the
-   * factory for all other attributes.
+   * Create a new ICUCollationKeyAnalyzer, using the specified collator.
+   *
    * @param collator CollationKey generator
    */
-  public ICUCollationAttributeFactory(Collator collator) {
-    this(TokenStream.DEFAULT_TOKEN_ATTRIBUTE_FACTORY, collator);
+  public ICUCollationKeyAnalyzer(Collator collator) {
+    this.factory = new ICUCollationAttributeFactory(collator);
   }
-  
-  /**
-   * Create an ICUCollationAttributeFactory, using the supplied Attribute 
-   * Factory as the factory for all other attributes.
-   * @param delegate Attribute Factory
-   * @param collator CollationKey generator
-   */
-  public ICUCollationAttributeFactory(AttributeFactory delegate, Collator collator) {
-    super(delegate, ICUCollatedTermAttributeImpl.class);
-    this.collator = collator;
-  }
-  
+
+
   @Override
-  public ICUCollatedTermAttributeImpl createInstance() {
-    return new ICUCollatedTermAttributeImpl(collator);
+  protected TokenStreamComponents createComponents(String fieldName) {
+    KeywordTokenizer tokenizer = new KeywordTokenizer(factory, KeywordTokenizer.DEFAULT_BUFFER_SIZE);
+    return new TokenStreamComponents(tokenizer, tokenizer);
   }
 }
