+++ lucene/analysis/icu/src/java/org/apache/lucene/analysis/icu/tokenattributes/ICUCollatedTermAttributeImpl.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.collation.tokenattributes;
+package org.apache.lucene.analysis.icu.tokenattributes;
 
 
 import org.apache.lucene.analysis.tokenattributes.CharTermAttributeImpl;
