+++ lucene/analysis/icu/src/test/org/apache/lucene/analysis/icu/TestICUCollationDocValuesField.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.collation;
+package org.apache.lucene.analysis.icu;
 
 
 import org.apache.lucene.document.Document;
